requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
anthropic
python-dotenv
httpx
uagents
